Moved Memory from core to memory
