+++ tests/tests_agents/tests_standard/test_numpy.py
@@ -9,7 +9,8 @@ from copy import deepcopy
 
 from learnrl.environments import RdNimEnv
 from learnrl.agents import StandardAgent
-from learnrl.core import Memory, Playground
+from learnrl.core import Playground
+from learnrl.memory import Memory
 from learnrl.evaluation import MonteCarlo, TemporalDifference, QLearning
 from learnrl.control import Greedy, Ucb, Puct
